requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Vision-Language Models
transformers>=4.36.0
open-clip-torch>=2.24.0
groundingdino-py>=0.4.0

# Object Detection
ultralytics>=8.1.0  # YOLOv8

# Drone Simulation
gymnasium>=0.29.0
pybullet>=3.2.5
# gym-pybullet-drones: install from source (see README)

# LLM APIs
openai>=1.12.0
anthropic>=0.18.0

# Communication
pyzmq>=25.1.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0
Pillow>=10.0.0
opencv-python>=4.8.0
pycocotools>=2.0.7

# Logging and experiment tracking
wandb>=0.16.0
jsonlines>=4.0.0
